Code formatting issue.

+++ src/main/java/com/maehem/javamancer/neuro/view/popup/DialogPopupPane.java
@@ -35,7 +35,7 @@ import com.maehem.javamancer.neuro.view.PopupListener;
  */
 public abstract class DialogPopupPane extends PopupPane {
 
-    private static final int WIDTH = (int) (624);
+    private static final int WIDTH = 624;
     private static final int HEIGHT = 70;
     private static final int X = 8;
     private static final int Y = 2;
